backend/requirements.txt
asgiref==3.8.1
Django==4.2.11
psycopg2==2.9.3
python-dateutil==2.8.2
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
coverage==6.3.2
gunicorn==20.1.0
ruff==0.3.1
django-extensions==3.2.3
pygraphviz==1.12
